Make the date format preference work again.
It is only used for numeric dates -- spelled-out dates have such a complex
variety of formats that they can only be meaningfully formatted from
locale strings.

In addition, the preference is left null when initializing, on the assumption
that the locale will still specify a more useful numeric format than we can
guess as part of a build-wide configuration.

But if the user has specified a format, the date will be formatted in the
order they asked for, with locale-appropriate punctuation substituted in.