Adds new attributes for specifying icons to be used in the Contacts app.
Adds the <icon> and <icon-default> tags and their associated attributes.
These tags should be used in meta-data associated with a sync adapter
that supplies contact data to the contacts provider. The <icon> tag
specifies an icon that is associated with a specific mimetype for that
package, and the <icon-default> tag specifies a default icon associated
with the sync adapter package.