Make the DatePicker respect the date format setting if the date is numeric.
In some locales, there are no abbreviated month names; the abbreviated
date formats are essentially numeric.  If the user is in such a locale,
have the DatePicker respect the date format setting so that the order
of the fields will match other numeric-only dates.

In locales that have abbreviated month names, continue to use the order
that is normal in spelled-out dates.

And update the order in updateDate() so that the new order is reflected
if you change the order setting and immediately go to change the date
without leaving and returning to the Date & Time settings in between.

At the same time, change DateFormat.getDateFormatOrder() back to working
the way it did in cupcake (prioritizing the date order preference over
the locale), even though the DatePicker no longer calls the method.

Bug 1805085